+++ packages/generate/_test/routes_static_imports.ts
@@ -3,12 +3,14 @@
 import { byMethod } from "@http/route/by-method";
 import { byPattern } from "@http/route/by-pattern";
 import { cascade } from "@http/route/cascade";
-import * as route_2 from "./routes/about.ts";
-import route_3 from "./routes/index.ts";
+import * as route_2 from "./handle_txt.ts";
+import * as route_3 from "./routes/about.ts";
+import route_4 from "./routes/index.ts";
 import * as route_1 from "./routes/user/:name/index.ts";
 
 export default cascade(
   byPattern("/user/:name{/}?", byMethod(route_1)),
-  byPattern("/about", byMethod(route_2)),
-  byPattern("/", route_3),
+  byPattern("/raw", byMethod(route_2)),
+  byPattern("/about", byMethod(route_3)),
+  byPattern("/", route_4),
 );