+++ packages/generate/_test/txt_route_mapper.ts
@@ -1,0 +1,10 @@
+import type { DiscoveredPath, DiscoveredRoute } from "@http/discovery/types";
+
+export default function txtRouteMapper({ ext, pattern }: DiscoveredPath): DiscoveredRoute[] {
+  return ext === ".txt"
+    ? [{
+        pattern,
+        module: import.meta.resolve("./handle_txt.ts"),
+      }]
+    : [];
+}